backend/pyproject.toml
[tool.poetry]
name = "retail-dss-backend"
version = "2.0.0"
description = "Retail Decision Support System Backend API"
authors = ["Ibrahim Turkyilmaz <admin@retaildss.com>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
asyncpg = "^0.31.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.27.0"
reportlab = "^4.0.0"
qrcode = "^7.4.2"
pandas = "^2.2.0"
numpy = "^1.26.0"
scikit-learn = "^1.4.0"
slowapi = "^0.1.9"
google-generativeai = "^0.4.0"
# Optional/Legacy
openpyxl = "^3.1.2"
matplotlib = "^3.8.0"
# Email
aiosmtplib = "^3.0.1" 

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.1.0"
isort = "^5.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
